requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
joblib==1.3.2
google-generativeai==0.3.1
Werkzeug==2.3.7